Fix expanding of multiple times nested groups on Special:LanguageStats
Imagine a scenario where A and B are aggregate groups and c is a normal
group. Moreover, A includes B and c, and B also includes c. Now, when
expanding A, we would show c twice on the stats page. After this patch,
we only show c once after expanding A, and twice when also B is expanded.

The reason for the previous behavior was that we store the parent group
of the group c in the output. Group c will be output twice, once with
parent being A, and second time with parent being B.

We cache the output of each row separately. However, the cache key did
not take into account that parent group can differ all else being equal.

Because of that, we always cached c having parent A (fortunately, having
it as B would have been even more confusing). The second time we reuse
the cache and the parent is now incorrectly A again.

Fixed by changing the cache key to take the parent into account.

Reported by Raymond on IRC for translatewiki.net, where
A = MediaWiki extensions
B = Blue Spice extensions
c = any Blue Spice extension

Change-Id: If059a47d0109f14b9ec26b27d979d7dc432b6c28