Fix skipping translation memory self-suggestions
When editing an already translated message, the translation memory
will return its translation together with any other matches. This
translation should be ignored and not shown as a suggestion. The
check for this was failing when the main translation memory was
configured to be public. In this case the code calls it remotely
via ApiTTMServer, which does not provide result attribute "local".
Fixed this via a slightly hacky URL comparison.

This was not enough to fix the issue, as the client side check for
the propery was done incorrectly. The query uses formatversion=1,
which means that boolean attributes are either missing (false) or
empty string (true). Empty string evaluate to false in a simple
truth check in JavaScript. Changed the comparison to be correct
and added future proofing for migration to formatversion=2.

This also includes slight rewrite of showTranslationMemory to use
less DOM manipulation and looping by storing already created
suggestions in an object with a counter.

Change-Id: I6dae7ad831f834de823984152143eea1b078f548