Add support to limit validations to certain messages
Added a new "keymatch" option to the VALIDATORS option when
specifying the group configuration.

Keymatch is an array with each option being either a string
or a prototype. The prototype has two properties - type and
pattern. Type can be regex or wildcard, and the pattern is
a string that will be used for matching.

Bug: T231705
Change-Id: I396e9ef5627073928b429284b5749de148570a72