Add SmartFormat plural validator
This is quite a bit simpler because it is a native inline plural
syntax which we only need to parse, not generate. For now it doesn't
support nesting, nor anything else that requires } as it will stop
parsing there. That can be improved in the future if needed.

I added insertables too. It would be nice if the interface passed
target language, so we could customize the insertable to match
the number of forms required in the target language.

Aside, the SmartFormat syntax has, imho, quite bad design flaws that
makes it impossible to parse the text properly without knowing types
of parameters, as the syntax is very highly ambiguous. Features like
named parameters, optional zero format, variables inside the plural
and the weird {} seen in examples are not supported.

Bug: T231181
Change-Id: Idf872e0ca54f48af21a124beb7a81d76237e044e