Fix incorrect translation restriction check
The condition needed to be reversed. Ended up changing the condition
to check for an empty array instead.

Another scenario: if a page was marked for translation without
priority languages, and priority languages were set after some non
priority languages received translations, we will continue to
allow users to edit the non priority translations. This is because
we recently updated the code to check for restrictions only during
translation creation.

Bug: T238099
Change-Id: I3561bee04991273bd217017ca58529d528db7e47