Handle error thrown when fetching content for unavailable message
Shifted to using RevisionStore rather than the deprecated Revision
when fetching content. Added a null check for the RevisionRecord
before trying to fetch the content. If the RevisionRecord is null,
simply setting the content for the message to be an empty string.

Added error handler around the main API business logic to log
necessary details and then rethrow the exception. This will be
helpful to debug other issues in the future.

Added an error handler for setRename JS function to display the user
an error message if something goes wrong.

Reproducing the issue: The only way I was able to reproduce this
issue is by deleting a non source message in the wiki, and then
renaming the source message key in the incoming changes. But I'm
not sure if this is the only way to reproduce it.

Bug: T238516
Change-Id: I8bbf1bf86764244c3e9afd15808adf7499466f7f