Enable "opt-out" translation aware transclusion for templates
Add BeforeParserFetchTemplateRevisionRecord for MW >= 1.36
This feature will not be supported for MW < 1.36.

Fetch the language of the source page where the template is
being included, and then fetch the RevisionRecord for the
transcluded template in that target language.

If the template has not been translated to that target language,
we fetch the RevisionRecord from the source language of the
template. In this case an additional dependency is added on the
source language of the template.

An option has been added when marking a page for translation
that allows translation admins to decide if a template should have
translation aware transclusion.

Existing pages that do not have this option set are treated as not
supporting transclusion, but this will be changed in the future.

Based On: Iec2bcc0417a16

Bug: T47096
Change-Id: I6d76b9ed1ac8efffbdb04670eb598182f9beb0ac
Depends-On: Ia5b5d339706ce4084c16948300e0e3418b11792e