MessageGroupCache: Use more efficient way of listing keys
This should use less space and memory, and should also be faster than
(un)serialize of big arrays.

This is backwards compatible, so I am not bumbing the version number.

Added a test that ensures they order of keys is retrained.

To make testing easier, I made the cache file path injectable, pending
future work of moving the fallback out of this class alltogether.

I also replaced once use of $group->getId() in hopes that in future
this class can depend on much narrower interface.

Change-Id: Id0486cdda31a63d8e8a3e957936ce6cec21be828