SpecialSupportedLanguages: simplify per language output
While investigating T248125 I found that the code is unnecessary
complex. It was throwing away translation edit count information
for anonymous users. In addition, it was using user_editcount
field instead of the actual number of edits in the requested
language. I think the latter number is more useful, so I scraped
the query for user_editcount. Last edit is now calculated using
the main query, where it shouldn't affect performance given it is
already scanning a lot of rows and picking up one column more is
fast.

This does mean that the numbers are more out of date, as time of
last edit and edit count was previously cached for one hour, while
now they are cached for one day. I think one day is still sufficient.

Change-Id: I8eb1ece37ae7711f38bfa2a3f7d42cd322775a8e