JsonFFS: Remove redundant author handling
It was re-parsing the source file to get authors from the file, but this
was already done by SimpleFFS::write through SimpleFFS::tryReadSource,
which calls MessageCollection::addCollectionAuthors.

Apart from doing unnecessary work, this also made it so that author
blacklist was not getting applied properly for authors present in the
source file. This prevented cleaning up old translator names from the
files after being renamed.

Also simplified the code using ??.

Change-Id: I44da9a19ef3f5e8beb9eb5649ba9de9a3e4c940b