export.php: Make FILES.targetPattern configurable
This is technically breaking change, since --offline-gettext-format
now requires a parameter. But since it was just recently added, I
think it is okay. It may be specified as --offline-gettext-format=
to give an empty pattern, which defaults to %GROUPID%/%CODE%.po
that was the hardcoded format earlier.

This adds a new path variable %GROUPID%, which will be documented in
https://www.mediawiki.org/wiki/Help:Extension:Translate/Group_configuration#pathvar
after merging.

Bug: T248200
Change-Id: I0a71097cf513639bd6d537a4751eaf76a0e7e9b7