Mangler: Don't remove prefix on unmangle if pattern doesn't match
If prefix for the mangler is datadump- and patterns is action-*,
during unmangle, not only will datadump prefix be removed from
messages matching action-*, but it will also be removed from any
message that has the prefix datadump-.

This patch fixes the issue so that unmangle only removes prefix from
matching patterns.

Also added type hints for StringMatcherTest file.

Change-Id: I8435be497f23818c1f7be25e484c6780e90d52f3