MovePage: Remove subpage option when checking if move is possible
When checking if a page move is possible, we list all pages including
the subpages. A notice, in bold, is shown for translatable subpages
that cannot be moved. If there are no subpages to be moved then a not
found message is displayed rather than displaying just the heading.

On the second step, the user can select if they would like to move
the subpages also.

Updated the count of "pages to be moved" message to show the number of
subpages that will be moved.

A similar change has been done for the DeletePage.

Bug: T168591
Related: I6cb7b735c69cd05204487939668598d3ac66e33f
Change-Id: I65523ec80d8e5b3ed3c27b8a19129d187e8a319f