DeletePage: Display subpages without strikethrough
When the user clicks on List pages for deletion, we now list all
pages, including the subpages without a strike-through. If there
are no pages to be deleted under a section, then a not found message
is displayed under that heading.

On the second step, the user can select if they would like to delete
the subpages also.

Updated the count of "pages to be deleted" message to list subpages
separately.

Similar changes are made in Translation move page.

Bug: T168591
Related: I65523ec80d8e5b3ed3c27b8a19129d187e8a319f
Change-Id: I6cb7b735c69cd05204487939668598d3ac66e33f