MessageCollection: Avoid notices in certain API queries
::rowToKey would complain with notices in certain kind of queries
where call to ::initMessages would be delayed in such a way that
keys would get filtered further after initial call of ::loadData
from a filter. Second condition is that the filter itself would
not match anything, because if it matches, it would call ::rowToKey
which would call ::getReverseMap and cache the results.

Easiest way to reproduce with any group is an API call like this:
  api.php?action=query&format=jsonfm&list=messagecollection
  &mcgroup=[valid group id]&mclanguage=[valid language]
  &mcfilter=!last-translator:-1&mclimit=1

Change-Id: If68b8a64fc42d3003c5cb96ddde64e044f8dcf40