MessageGroupStats: Avoid COMMIT while atomic section is open
This commit partially reverts d5b96336fa0 by re-adding a flag that
determines whether it is safe to write updates immediately.

This issue happens when a change, like an edit, inside a job in the
job queue trigger a message group stat update. Since the check
whether we are in job runner passes, it assumed it can write updates
immediately. But it is not safe while running inside a "SaveComplete"
hook handler, because it opens a transaction which we cannot flush.

Differences to that commit are that now the flag is set in the job
and only if it is deemed safe. To be deemed safe, we require that
the job is run in the JobQueue as a separate job.

In theory, this issue could still be triggered if the following
conditions are met:
* a job in the JobQueue makes an edit,
* the edit triggers a "SaveComplete" hook which creates a rebuild job,
* and the job is run immediately, not as separate Job.

Looking at the current code, this job is never executed immediately,
so this code is safe for now.

Bug: T247809
Change-Id: I1c42cbfcea0d73d4f91facdacc4f21fa6ed9ed61