Fix "nothing to translate" error on moving from review to page mode
Nothing to translate message was being displayed when moving from
proofread mode to page mode. When the page mode is selected and the
filter has !last-translator, we are clearing the messages and
setting the default filter for translate mode. In this case the
messagetable is initialized but without any messages, which causes
the "nothing to translate" message to be displayed.

Add an additional check to ensure that the messagetable is currently
not in the loading state before displaying the message.

Noticed another issue where moving from untranslated to review
displays an empty messagetable without any messages or notices.

Bug: T49986
Change-Id: I4028c122653e6858b50b7a43a9ed1791efb9b686