Display help message when editing a translatable page source
Add a help message that will be displayed to users when they edit
a translatable page source.

In addition when user is in VE visual mode, display a warning dialog
to the user stating that editing in visual mode is not supported.

Note that we do not shift the VE mode from visual to source.

Bug: T192052
Change-Id: I03ea64353b2d5b28d684f90936cab75ea4fd40d2