Do not query TTMServer mirrors
The querying code was not updated when support for mirrors were added.
Due to this, we were queuing the mirrors as well, serially, unless they
were public (not in case of WMF). This doubled the query time for no
benefit.

Code was refactored for clarity and to consolidate and unify the logic of
determining what can be queried over HTTP (through MediaWiki Action API).

Bug: T249906
Change-Id: I09ae7ded332d203cfd5df01c316bc3cf2d703c40