Add new MessageValidator interface
The name is not ideal, as it clashes with the MessageValidator class
name. But I am expecting one more update when we add a MessageRecord
class, so we can try to find a better name then.

Added an adapter for transparent support for classes using the old
interface. The classes can be now converted separately to the new
interface.

Assorted improvements to the touched code, including type hints and
renames to reduce the use of the term blacklist.

Test validators were updated to the new interface.

Change-Id: Id30185d3692fa3ec0ee085a9352380882b0f53bc