Rewrite BraceBalanceValidator
Convert to the new interface, and simplify code.

ValidationHelper::checkStringCountEqual was folded into this class,
since it was only used here.

Added strict types, since there was a bug that didn't get immediately
exposed because of lax type handling in PHP. Probably a good idea to
start using it more in new code.

Converted its test to a unit test.

Change-Id: Ie83273d0fd96d8be22c25dcb3f6ac1c2ff6a4557