Convert some page translation integration tests to unit test
TPParseTest is converted to a unit test, and the tests in it that
test TranslatablePage are moved to TranslatablePageTest.

Unit test moved away from TranslatablePageTest. One test remains,
but the class is renamed to TranslatablePageIntegration test to
avoid class name conflicts.

TPSectionTest was moved directly with some small formatting changes.

Small refactorings to TranslatablePage to avoid accessing services
during unit tests. This should be refactored more later to clean it
up better. Unused ::setRevision method was removed.

Change-Id: Ic5f34c505788d504a07ad4d1d2ebbc3cab99279c