SearchTranslations: Fix display of message groups in facet sidebar
This was broken during Ic7242276e2c3328d83a2996a14344ddb2f636777

Earlier mw.translate.findGroup was being called with an array and also
an object (from ext.translate.special.searchtranslations)

In the referenced patch, $.each was replaced with forEach, but forEach
does not work with objects. Another caveat is that forEach does not
break on return false.

Added a separate method in ext.translate.special.searchtranslations to
handle finding a group when groups is an object, using the "for in"
loop.

Replaced forEach with Array.some in ext.translate.base in order to
return early and stop the loop.

Bug: T257268
Change-Id: Ibe774e7808c67ce5782afe8322de787295435526