Reduce styles in Special:Translations
Most of the row classes were already unused. Now inheriting default
styles from wikitable, which allowed to remove everything except the
row coloring for outdated messages. I kept that and prefixed the class
name to avoid name collisions.

There are some changes in the display, but I didn't see them as very
important to keep, but can be restored if anyone disagrees. Mainly:
* table is no longer forced to full width

Change-Id: I8e378a60b2ba91b1257e71e1ad166cf8d2020c9a