Add table and aria-label for the translation stats graph
Display table containing data under the graph so that screen
readers can read this data out. The table is hidden by default,
but displayed when it gains focus.

Add role="img" to the graph. Make the graph focusable and add
content inside it to point to the table containing the textual
version of the data.

Bug: T263543
Change-Id: I8cf517a1c5d891b75242767130a9a37ed3217a70