Stop hiding page heading on pages other than Special:Translate
There is no good justification for hiding the header on these pages
* Special:ManageMessageGroups
* Special:AggregateGroups
* Special:TranslationStats
* Special:SupportedLanguages
* Special:LanguageStats
* Special:MessageGroupStats

Only Special:Translate has such a dedicated interface that there
can be a case for hiding the heading there.

Hiding the heading causes an issue with the Timeless skin that the
help page link on the indicator area is not clickable (if present).

Special:Translate does not have a help link currently.

Fixed by moving the hiding from the tabgroup module to styles for
Special:Translate.

Also fixed an issue where Special:ExportTranslations was part of a
tab group but was not itself showing the tab group.

Change-Id: Ie27cfd707cbd6b290395d36bd4845081cac155d0