Update cache version in group loaders to trigger recache
In Ifecb3345f1b7e15c7a497c8860e915d18d8b6ba8 we updated the
configuration for certain groups to use array based configuration
for Insertables and added a check to ensure non-array based
configuration is not supported.

Since configuration is cached, the new configuration changes made
are not picked up causing exceptions to be thrown for the old
array based Insertable configuration.

Update the cache version to ensure configurations are recached.

Change-Id: I5726cf956f1c45d32c9f2f60f0b26649bc998882