Allow importing empty content from external sources
In some cases it is important that we properly preserve empty content.
There was a check that skipped empty definitions and translations
This check was mostly used to mask various cases where empty content
was a mistake, such as being a placeholder for a missing translation.
In the past, there was also an issue that it was not possible
to create empty pages, but now there is no problem if it is done via
MessageUpdateJob. The translation UI does not allow saving empty
translations, but this shouldn't be a problem since these kind of
messages are marked as ignored.

Added warnings to Special:ManageMessageGroups for when the content
is empty to make it easier to detect these and different from other
cases where the diff may not be shown (e.g. some difference in white
space may not be detected).

Bug: T271294
Change-Id: I19b7c5a760d61e3034e928bcd6bceadfc61f6266