Do not lock pages indefinitely during translatable page moves
The moves can fail, in which case the lock would not be released.
This blocks, among other things, editing the pages. Add a timeout
so that the locks expire by itself after a while in case of a
failure. The duration was chosen by a gut feeling.

Related to task T206464

Change-Id: I425aed541fb44fced779fa87b0ed4e9cd12c5bdd