Add code to track groups with errors and failed messages
Failed messages and groups are tracked separately in the group
synchronization cache. This will allow issues with each messages
to be tracked and marked as resolved individually.

On the Special:ManageGroups we display the groups that have had
errors during the group synchronization process.

Each message that has failures is displayed and admins can view
the history of the messages.

In a subsequent patch, admins will be allowed to mark each failed
message as resolved. Once all are resolved the group is removed
from errors list.

To test this, stop automatic running of job queue and reduce the
$timeoutSeconds in GroupSynchronizationCache. Import some message
changes and then run completeExternalTranslation script.

Bug: T273671
Change-Id: Id667c72ada360fe4de0db835bf1c054242ca0db5