Skip importing other languages if source language is not safe
When running safe imports we would check each language under a
group individually and then if it only had additions we would
import it. This could lead to situations where translations would
get imported, but the source language message would not be
imported.

Update the safe check, to first consider the source language. If
the source language is safe to import, then proceed to import the
other languages otherwise skip the entire group.

Change-Id: I7b587a1b8a1e61fc0022f661b58294ee55fdca84