Do not enable supportsTransclusion by default
Previously, If an old page did not have a value set for supports
template transclusion, we still have the checkbox checked when the
translation admin is marking the page for translation.

Reason behind the initial decision: If supports template transclusion
was defaulted to false, later even if we flipped the default flag
to true, the pages marked as false would not support template
transclusion. This would leave a lot of page that could support
template transclusion with the supportsTransclusion flag as false.

Since this behavior may lead to accidentally breaking things, update
this behavior to:
* If an existing page is being marked for translation, and does not
  have the supports template transclusion flag set, do not select
  the checkbox
* If a new page is being marked for translation, have the checkbox
  checked by default.

Bug: T47096
Change-Id: I68c34f79132026f4ffd9924cf48c30b5a5c7327f