Rename Special:SupportedLanguages to Special:ActiveLanguages externally
This name better matches the page's contents and does not give a false
impressions that languages not listed here are not supported.

Internally the page is still identified by SupportedLanguages, changing
that would complicate this patch and may not be worth it at all.

The class was also moved to a namespace to avoid doing multiple renames.

As usual, PHPStorm could not resist making some code formatting changes.

Change-Id: I63b5131c6278dcdf51c9cae8c6641804915e7f3f