MessageGroupStats: Remove use of by-ref $updates variable
It seems that the only effect of passing this around by reference, is
that later evaluations of $updates act as if they are evaluating
the latest/current self::$updates. If that is true, then we might
as well do that instead without the added risk/confusion of a second
way of holding and referencing that array.

Bug: T209817
Change-Id: I4840b052696705ad68b1f6c67c635b2393acee02