Special:TranslationStats: Allow yearly stats
Had to replace code using array_splice, which did not preserve
integer-looking array keys.

Removed join to actor table, we are not interested in the user names,
unique identifier is enough.

Increased maximum days for queries from 200 to 400 to allow showing a
full year at once. Canary testing indicates that performance is
acceptable. Note: RegistrationStats query has a different limit of 1000
days, which was not changed.

Bug: T281244
Change-Id: I8a00b388e06ace01eda77de8ac1225dd24568fc9