Make TranslatablePageMessageGroupStore work with ONLY_FULL_GROUP_BY
Maybe it would work on MySQL 5.7+, which is documented as supporting
functional dependence, i.e. grouping by a primary key implies the
uniqueness of the rest of the row. But CI uses MariaDB which
does not claim to support it.

Bug: T281329
Change-Id: Ib12d640f7ac3e29ad1fc305d21eed3c3125f5047
(cherry picked from commit 7bdb760da14113fdc1330b4174c508d9db2b8adf)