Fix bug where last unit of the scale could be ignored
This is very obvious with the yearly scale. If your range includes
days from the last year (i.e.. days input is more the days than the
current number of days passed in the current year), then only results
from the past year are shown. It also incorrect has the asterisk to
signify stats are not complete. This happens because our increment
jumps from the past year to future of the current year.

Bug: T281244
Change-Id: Ib760dff3b0334a84674b3014f6b76261a7b7bcd0