InOtherLanguagesAid: Don't unconditionally remove 'en' from the list
This probably came from the assumption of English always being the default
language, and was supposed to prevent the source text from showing up as an
aid. That exact functionality was then reimplemented in 0b3e8aa, in a more
correct way, which made this line redundant. Keeping English translations is
actually quite valuable in cases where the source language isn't English.

Change-Id: I08f543e64c4f1205e1ad9a8f81cc62e1b59d83b6