ext.translate.base: Remove redundant 'pageshow' handler
This appears to be copied from core's mediawiki.confirmCloseWindow.js,
but without the relevant logic that made it useful (e.g. remove
the 'onbeforeunload' handler on pagehide, and then restore it on
pageshow). However, even that logic hasn't worked in core since 2013
and was only there to workaround a bug in Firefox 3.

While at it, also remove the `.translate` jQuery event namespace,
which I guess was also copied from there, but not actually used for
anything.

Not strictly needed for T283193, but tagging it since I cleaned this
up as part of code searching for that.

Bug: T283193
Change-Id: I8633a626c57382375ec7676e7bf4f5ed2916805c