Try to use WikiPage::doUserEditContent() instead of ::doEditContent()
::doUserEditContent() is available since 1.36 as a replacement
for ::doEditContent(), which has been deprecated. But, this
extension needs to keep supporting MediaWiki 1.35. So,
add a helper method to TranslateUtils that will call the
correct WikiPage method by checking if the WikiPage object
has a `doUserEditContent` method, and make use of that helper.

Results in passing a user where previously the fallback
to $wgUser was being used.

Bug: T255507
Change-Id: I1bbe9f4f5f0a1019f874a003d5e77c69ff77de07