MessageGroupCache: Optimize glob usage
When profiling with xdebug, I noticed that there were as many calls
to glob() as calls to this method. Turns out $uniqueId was too unique
and the glob caching did not work. Fixed this be using the $pattern
itself as the cache key. Before the fix this accounted for 20% of the
total time used.

Change-Id: Icb9419bf08d124fe59fc5c0e4a6a7c4331665116