BREAKING CHANGE: Remove TranslateMessageGroupPathVariables hook
When profiling with xdebug, FileBasedMessageGroup::replaceVariables
accounted for 50% of timed used by processMessageChanges.php before this
and the parent fix. The reason is that running hooks in a tight loop is
very expensive. This could be mitigated with caching, but according to
the code search there are no uses of this hook. For performance, it is
better to just hard-code more variables here if needed than make it
extensible.

Change-Id: I179907886be3fd840f08f0c0291681fa2426a70a