Fix fatal on Special:ImportTranslations during import
TypeError: Argument 2 passed to WikiPage::doUserEditContent() must
implement interface MediaWiki\Permissions\Authority, null given.

This is probably a regression introduced in TranslateUtils::doPageEdit
which calls WikiPage::doUserEditContent which no longer accepts null
for performer.

This is a minimal change to pass the performer properly.
The MessageWebImporter class could use with a lot of cleanup to remove
unused code.

Change-Id: Id504c5317f74d214dd6d970ecf9b4d1071414002