ExportTranslationsMaintenanceScript: Add --always-export-languages
Currently repong needs to invoke export twice for each group. This is
not ideal since any error output (like groups in sync) will be displayed
twice.

TranslateUtils::parseLanguageCodes was moved to the class, which was
the only one using it, and rewritten to avoid deprecated code.

Small tweaks to variable names and parameter descriptions. --skip was
renamed to --never-export-languages for clarity with BC.

Bug: T287476
Change-Id: I4fd574b37ce4eaed60954bd98e1da9ea71e46b5f