TranslatablePage: Remove getParse
Moved it to SpecialPageTranslation, which is the only user of it,
and made it private.

Also introduced TranslatablePage::DISPLAY_TITLE_UNIT_ID to replace
the hard coded unit name in different places.

This temporarily removes caching from the parses. Subsequent commits
will make it so that there is only one parse which is passed around.

Parsing is not expected to be a bottleneck that would cause problems
during the interim period before the subsequent commits are merged.

Change-Id: I37286adf0a471c65d2c07fa345b751ab0f51fd19