ExportTranslationsMaintenanceScript: Avoid creating empty files
The addition of --always-export-languages made it so that now they
are exported even if there are no translations. In Do-What-I-Mean
sense, avoid doing that by first checking if there are any
translations. This logic only applies if an threshold is specified.

Possibly the logic could be simplified, but I left it this way for
now to leave it possible to create empty files for some corner case
if that's really needed.

Bug: T287476
Change-Id: I29c6a937d4366fdf35fc3d512be6907f7c76c9f8