Special:PageTranslation: Avoid getting messages inside a loop
According to quick profile on meta.wikimedia.org, this page
constructs over 60k message objects. The contents of these strings
do not change, so we can cache them. A few list-type messages
with varying content are cached with a unique key or by inlining
the function in a way that allows caching.

Change-Id: I64f159973b8782518f85be31aee00a1e53d25152