Breaking change: Remove translation aids customization
In order to make it easier to implement the MessageGroup interface,
this patch removes getTranslationAids method from the interface.
According to code search, nothing is using this customization, so it
is just needless indirection. The TranslateTranslationAids hook is
also removed, since it no longer makes sense, and it is unused, too.

Change-Id: I94dc814a9d3dee3927d1e94ad5cecf2f00f9127b