Optimize queries on Special:LanguageStats
1) Use LinkBatch to eliminate page table queries
2) Avoid loading all of translate_metadata

For the latter, I removed the Translate:MessageGroupStats:isIncluded
hook, as there were no external callers and it is just overhead to
use a hook for this. I replaced it with a process cached query
that only returns rows for groups which have priority language
enforcement turned on. This replaced the preload call which was
just loading everything from translate_metadata, returning >30k
rows on metawiki. The new query also does a table scan, but returns
fewer rows. That shouldn't be a problem, especially if we can get
the empty rows removed from the table. MessageGroups::getPriority
is already cached.

Notes for testing: due to multiple layers of caching, testing this
patch is difficult. It is likely helpful to change CACHE_ANYTHING
in SpecialLanguageStats to 'hash' for the duration of testing.

Bug: T204026
Change-Id: Iaaf74ade2574793857e7997d2aa3836ed055e833