PageTranslation: Prevent non-wikitext pages from being marked for translation
This change prevents pages without a Wikitext content model from being
marked for translation. Even if translation markup is placed on the
page, the "Mark this page for translation" header won't be shown and the
page won't appear in Special:PageTranslation.

Non-wikitext pages that are already marked for translation will still
appear in Special:PageTranslation (including pages with
currently-pending changes). If the page is edited, it will be listed as
"broken" and can't be re-marked for translation. Translation of
already-marked pages may still be possible until they are removed by a
translation administrator.

Bug: T163254
Change-Id: I81c5064062fb06ca0e775626932b1f5412dfe03d