Backports: Preserve order of message keys
Previously, any new translations picked from the source branch (e.g.
main) to the target (stable) branch were appended at the end. Now
they are kept in the order as they appear in the source language
file in the target branch.

This will also re-order keys if they were not in order already AND
there was some other update.

Bug: T272830
Change-Id: Ifb6b055c5ec7a423a5ead66717cdee0aae2155f2