Optimize metadata query on Special:Translate
The message group selector only needs to show message groups that
are available for the given target language. Add a new parameter
to meta=messagegroups API that allows doing that filtering on the
server side using existing optimized filtering query.

Metadata is now loaded only if prioritylangs or priorityforce prop
is selected. They are not by default, and the message group selector
no longer requests them, as they was replaced with the server side
filtering.

This also fixes unrelated existing issue, that the message group stats
in the selector were not properly updated for the new language after
a language change.

Bug: T204026
Change-Id: I8e832a45d2233317f03c22708509fb8bf2e1f96f