BackportTranslationsMaintenanceScript: reduce noise in output
When running autobackport-mediawiki, each group will print a bunch
of lines to error log about skipping aggregate groups that match
the given pattern (e.g. ext-*). This information is not useful, so
do not print that warning if --filter-path is given, which indicates
that we are only interested in specific groups anyway.

Bug: T272830
Change-Id: Ie476d131e22389297d425406d563442bb5f01839