JsonFFS: Fix unmangling of message keys during backports
Backports use readFromVariable and generateFile. Former does mangling
while unmangling was only done in writeReal (which calls generateFile).

Simplest fix is to move unmangling from writeReal to generateFile. This
mangling is not really needed for backports, but it is not easy to skip
it for now. That can be done after a refactoring where mangling is no
longer the responsibility of file format handlers.

Bug: T272830
Change-Id: Ib06677de79d654275a51d5e366f80d0284222c2b