Rename TranslatablePageMoveJob to TranslatableBundleMoveJob
This class will also be used to move MessageBundle. Translatable page
and Message bundles are together referred to as Translatable bundles.

Keep the old job reference incase there are existing jobs in the
queue.

Testing report: If an old job is in the queue, and this patch is
deployed, the move works fine.

Bug: T300496
Change-Id: I82001726c4be499e7643b5f99e0a418face89c18