Avoid writing to objectcache table
This is the last place in production that has entries in objectcache
table. For cases like this, you should use mainstash which is currently
on redis but will move to its own dedicated database system in the
future. It has all the benefits of cross-dc storage and no downside.

This is blocking the schema change planned on objectcache table.

Bug: T272512
Change-Id: Ieeb3a87e4169b33d0bc9dcfcdb762af996d74469