Add support to move Message bundles
Add a message bundle source page cache in order to perform lookups
quickly.

Convert TranslatableBundle to an abstract class as some methods that
are used by TranslatablePage and MessageBundles are now present in
this class.

Update display messages and logging to account for the addition of
the new MessageBundle support.

Rename mb:ready tag to mb:valid since Message bundles don't have a
marking step.

One important thing to note is that, after moving a MessageBundle we
have to add the "mb:valid" tag to the latest revision of the bundle.

Bug: T300496
Change-Id: I1e596d6bf9300bf80f5406aa337b52cccc1c9ec9